Rename setup module to helper
